+++ dev/src/route.js
@@ -94,32 +94,7 @@
         },
 
         interpolate : function(replacements) {
-            if (typeof this._pattern !== 'string') {
-                throw new Error('Route pattern should be a string.');
-            }
-            var replaceFn = function(match, prop){
-                    if (prop in replacements) {
-                        return replacements[prop];
-                    } else if (match.indexOf('{') !== -1){
-                        throw new Error('The segment '+ match +' is required.');
-                    } else {
-                        return '';
-                    }
-                },
-                str;
-
-            //TODO: extract this logic into pattern lexer and reuse TOKENS
-            str = this._pattern
-                        .replace(/([:}]|\w(?=\/))\/?(:)/g, '$1__CR_OS__$2')
-                        .replace(/\{([^}*]+)\*?\}/g, replaceFn)
-                        .replace(/:([^:*]+)\*?:/g, replaceFn)
-                        .replace(/(?:__CR_OS__)+$/, '') // remove trailing
-                        .replace(/__CR_OS__/g, '/'); // add slash between segments
-
-            if (! this._matchRegexp.test(str) ) {
-                throw new Error('The generated string "'+ str +'" doesn\'t match the pattern "'+ this._pattern +'". Check supplied arguments.');
-            }
-            return str;
+            return crossroads.patternLexer.interpolate(this._pattern, replacements);
         },
 
         dispose : function () {